third_party/blink/web_tests/external/wpt/css/css-shadow-parts/interaction-with-tree-abiding-expected.txt

This is a testharness.js-based test.
PASS ::before in selected host is styled
PASS ::after in selected host is styled
FAIL ::placeholder in selected host is styled assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
